Move probablistic and tail sampling processors under common parent
Phase 1 change to defining sampling processors. This structure follows
what is done in the receiver and exporter folders.
